Refactor mapstruct
