--- src/main/java/com/teethcare/config/SecurityConfig.java
@@ -1,27 +1,0 @@
-//package com.teethcare.config;
-//
-//import lombok.RequiredArgsConstructor;
-//import org.springframework.context.annotation.Configuration;
-////import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-////import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-////import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-////import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-////import org.springframework.security.core.userdetails.UserDetailsService;
-////import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-//
-//@Configuration
-////@EnableWebSecurity
-//@RequiredArgsConstructor
-//public class SecurityConfig extends WebSecurityConfigurerAdapter {
-////    private  final UserDetailsService userDetailsService;
-////    private final BCryptPasswordEncoder bCryptPasswordEncoder;
-////    @Override
-////    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-////        auth.userDetailsService(userDetailsService).passwordEncoder(bCryptPasswordEncoder);
-////    }
-////
-////    @Override
-////    protected void configure(HttpSecurity http) throws Exception {
-////        super.configure(http);
-////    }
-//}
