+++ src/main/java/com/teethcare/config/mapper/MapStructMapper.java
@@ -1,0 +1,33 @@
+package com.teethcare.config.mapper;
+
+import com.teethcare.model.request.ClinicRequest;
+import com.teethcare.model.entity.Account;
+import com.teethcare.model.entity.Clinic;
+import com.teethcare.model.entity.CustomerService;
+import com.teethcare.model.entity.Dentist;
+import com.teethcare.model.response.AccountResponse;
+import com.teethcare.model.response.CustomerServiceResponse;
+import com.teethcare.model.response.DentistResponse;
+import org.mapstruct.*;
+
+import java.util.List;
+
+//@Mapper(componentModel = "spring")
+public interface MapStructMapper {
+//    @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
+//    @Mapping(target = "managerId", ignore = true)
+//    @Mapping(target = "locationId", ignore = true)
+//    @Mapping(target = "avgRatingScore", ignore = true)
+//    @Mapping(target = "taxCode", ignore = true)
+//    @Mapping(target = "status", ignore = true)
+//    void updateClinicFromDTO(ClinicRequest dto, @MappingTarget Clinic entity);
+
+//    @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
+//    @Mapping(source = "dateOfBirth", target = "dateOfBirth", dateFormat = "MM/dd/yyyy")
+//    AccountResponse mapAccountToAccountDTO(Account account);
+//
+////    @Mapping(source = "dateOfBirth", target = "dateOfBirth", dateFormat = "MM/dd/yyyy")
+//    List<AccountResponse> mapAccountListToAccountDTOList(List<Account> account);
+
+
+}
