--- src/main/java/com/teethcare/controller/AuthController.java
@@ -1,4 +1,0 @@
-package com.teethcare.controller;
-
-public class AuthController {
-}
