+++ src/main/java/com/teethcare/model/entity/CustomerService.java
@@ -9,25 +9,18 @@ import lombok.Setter;
 
 import javax.persistence.*;
 
+@Getter
+@Setter
 @NoArgsConstructor
 @Entity
 @Table(name = "customer_service")
 @PrimaryKeyJoinColumn(name="account_id")
 public class CustomerService extends Account {
 
-//    @ManyToOne(cascade = {CascadeType.DETACH, CascadeType.MERGE,
-//            CascadeType.PERSIST, CascadeType.REFRESH})
-//    @JoinColumn(name = "clinic_id")
-    @ManyToOne(fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY,
+            cascade = {CascadeType.DETACH, CascadeType.MERGE,
+            CascadeType.PERSIST, CascadeType.REFRESH})
     @JoinColumn(name = "clinic_id")
     @JsonManagedReference
     private Clinic clinic;
-
-    public Clinic getClinic() {
-        return clinic;
-    }
-
-    public void setClinic(Clinic clinic) {
-        this.clinic = clinic;
-    }
 }
