+++ src/main/java/com/teethcare/model/entity/Dentist.java
@@ -1,6 +1,5 @@
 package com.teethcare.model.entity;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,6 +8,8 @@ import lombok.Setter;
 import javax.persistence.*;
 
 @NoArgsConstructor
+@Getter
+@Setter
 @Entity
 @Table(name = "dentist")
 @PrimaryKeyJoinColumn(name="account_id")
@@ -24,28 +25,4 @@ public class Dentist extends Account {
     @JoinColumn(name = "clinic_id")
     @JsonManagedReference
     private Clinic clinic;
-
-    public String getSpecialization() {
-        return specialization;
-    }
-
-    public void setSpecialization(String specialization) {
-        this.specialization = specialization;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Clinic getClinic() {
-        return clinic;
-    }
-
-    public void setClinic(Clinic clinic) {
-        this.clinic = clinic;
-    }
 }
