--- src/main/java/com/teethcare/model/request/LoginRequest.java
@@ -1,4 +1,0 @@
-package com.teethcare.model.request;
-
-public class LoginRequest {
-}
